requirements.txt
confluent-kafka==2.3.0
mysql-connector-python==8.3.0
PyYAML==6.0.1
mysqlclient==2.2.3
pandas==2.2.1
pymongo==4.6.2
cassandra-driver==3.29.1